Authenticate users when sending them Studio and the Editor
This requires opencast/opencast#4412 and an appropriately configured
Opencast, obviously. This also makes a proper button out of the link.
Semantically, this is not a static link anymore. We have to generate
the target on interaction. This breaks for example with "open in
a new tab", so we just take away those possibilities entirely.